# public/api-docs/swagger.yaml
swagger: "2.0"
host: festivals-app.herokuapp.com
basePath: /api
info:
  version: 1.0.0
  title: festivals
...
schemes:
  - https
produces:
  - application/vnd.festivals.v1+json
  - application/vnd.festivals.v1+xml
security:
  - oauth2token:
    - public
paths:
...
  "/festivals/{id}/categories":
    get:
      tags:
        - categories
      summary: Get festival categories
      description: Get festival categories collection
      operationId: getFestivalCategories
      produces:
        - application/vnd.festivals.v1+json
      security:
        - oauth2token:
          - categories:get
      parameters:
        - name: id
          in: path
          description: Festival id
          required: true
          type: string
        - name: updatedAt
          description: updatedAt
          required: false
          type: string
          in: query
        - name: parent.id
          description: category parent id
          required: false
          type: string
          in: query
        - name: name
          description: name
          required: false
          type: string
          in: query
        - name: limit
          description: limit
          required: false
          type: integer
          in: query
          default: 100
        - name: offset
          description: offset
          required: false
          type: integer
          in: query
      responses:
        "200":
          description: successful operation
          schema:
            $ref: "#/definitions/CategoriesCollectionResponse"
        "404":
          description: Festival not found
    post:
      tags:
        - categories
      summary: Create festival category
      description: Create festival category
      operationId: createFestivalCategory
      consumes:
        - application/vnd.festivals.v1+json
      produces:
        - application/vnd.festivals.v1+json
      security:
        - oauth2token:
          - categories:create
      parameters:
        - name: id
          in: path
          description: Festival id
          required: true
          type: string
        - in: body
          name: festivalCategoryRequest
          description: Festival category object
          required: true
          schema:
            $ref: "#/definitions/CategoryRequest"
      responses:
        "201":
          description: successful operation
          schema:
            $ref: "#/definitions/CategoryResponse"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Festival not found
        "406":
          description: The request you made is not acceptable
        "422":
          description: Invalid parameter value
  "/festivals/{id}/categories/{category.id}":
    get:
      tags:
        - categories
      summary: Get festival category
      description: Get festival category
      operationId: getFestivalCategory
      produces:
        - application/vnd.festivals.v1+json
      security:
        - oauth2token:
          - categories:get
      parameters:
        - name: id
          in: path
          description: Festival id
          required: true
          type: string
        - name: category.id
          in: path
          description: Category id
          required: true
          type: string
      responses:
        "200":
          description: successful operation
          schema:
            $ref: "#/definitions/CategoryResponse"
        "404":
          description: Not found
    delete:
      tags:
        - categories
      summary: Delete festival category
      description: Delete festival category
      operationId: deleteFestivalCategory
      security:
        - oauth2token:
          - categories:delete
      parameters:
        - name: id
          in: path
          description: Festival id
          required: true
          type: string
        - name: category.id
          in: path
          description: Category id
          required: true
          type: string
      responses:
        "204":
          description: successful operation
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    put:
      tags:
        - categories
      summary: Update festival category
      description: Update festival category
      operationId: updateFestivalCategory
      consumes:
        - application/vnd.festivals.v1+json
      produces:
        - application/vnd.festivals.v1+json
      security:
        - oauth2token:
          - categories:update
      parameters:
        - name: id
          in: path
          description: Festival id
          required: true
          type: string
        - name: category.id
          in: path
          description: Category id
          required: true
          type: string
        - in: body
          name: festivalCategoryRequest
          description: Festival category object
          required: true
          schema:
            $ref: "#/definitions/CategoryRequest"
      responses:
        "200":
          description: successful operation
          schema:
            $ref: "#/definitions/CategoryResponse"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
        "406":
          description: The request you made is not acceptable
        "422":
          description: Invalid parameter value
...
securityDefinitions:
  oauth2token:
    type: oauth2
    flow: implicit
    authorizationUrl: https://vehicle-history.io/oauth/authorization
    scopes:
      user: user auth scope
definitions:
...
  CategoryRequest:
    properties:
      parent:
        type: string
        description: Festival category parent id
      name:
        type: string
        description: Festival category name
    description: Festival category request
    required:
      - parent
      - name
  CategoryResponse:
    properties:
      id:
        type: string
        description: Category id
      name:
        type: string
        description: Category name
      parent:
        type: string
        description: Category parent
      createdAt:
        type: string
        description: Category createdAt
      updatedAt:
        type: string
        description: Category updatedAt
    description: Festival category response
    required:
      - id
      - name
      - parent
      - createdAt
      - updatedAt
...
  CategoriesCollectionResponse:
    properties:
      total:
        type: integer
        description: Total festival categories
      categories:
        type: array
        items:
          $ref: "#/definitions/CategoryResponse"
    description: Festival categories collection
    required:
      - total
      - categories
...